requirements.txt
# 필수 라이브러리

# 데이터 처리
pandas>=1.5.0
numpy>=1.21.0

# 웹 앱
streamlit>=1.28.0

# 시각화
plotly>=5.15.0

# API 클라이언트
requests>=2.28.0

# 환경변수 관리
python-dotenv>=1.0.0

# LLM 연동 (GPT)
openai>=1.0.0

# 웹 스크래핑 및 파싱
beautifulsoup4>=4.11.0
lxml>=4.9.0
requests-html>=0.10.0
chardet>=5.0.0

# 한국 주식 데이터
pykrx>=1.0.0

# Jupyter 환경 (IPython)
ipython>=8.0.0

# 동시성 처리 (캐시 생성용)
# concurrent.futures는 Python 표준 라이브러리이므로 별도 설치 불필요